+++ server/runtime_javascript_init.go
@@ -69,6 +69,7 @@ type RuntimeJavascriptCallbacks struct {
 	TournamentEnd                  string
 	TournamentReset                string
 	LeaderboardReset               string
+	Shutdown                       string
 	PurchaseNotificationApple      string
 	SubscriptionNotificationApple  string
 	PurchaseNotificationGoogle     string
@@ -104,6 +105,7 @@ func (im *RuntimeJavascriptInitModule) mappings(r *goja.Runtime) map[string]func
 		"registerTournamentEnd":                           im.registerTournamentEnd(r),
 		"registerTournamentReset":                         im.registerTournamentReset(r),
 		"registerLeaderboardReset":                        im.registerLeaderboardReset(r),
+		"registerShutdown":                                im.registerShutdown(r),
 		"registerPurchaseNotificationApple":               im.registerPurchaseNotificationApple(r),
 		"registerSubscriptionNotificationApple":           im.registerSubscriptionNotificationApple(r),
 		"registerPurchaseNotificationGoogle":              im.registerPurchaseNotificationGoogle(r),
@@ -1085,6 +1087,10 @@ func (im *RuntimeJavascriptInitModule) registerHook(r *goja.Runtime, execMode Ru
 		im.registerCallbackFn(execMode, lKey, fnKey)
 		im.announceCallbackFn(execMode, lKey)
 
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
+
 		return goja.Undefined()
 	}
 }
@@ -1187,6 +1193,10 @@ func (im *RuntimeJavascriptInitModule) registerRtBefore(r *goja.Runtime) func(go
 		im.registerCallbackFn(RuntimeExecutionModeBefore, lKey, fnKey)
 		im.announceCallbackFn(RuntimeExecutionModeBefore, lKey)
 
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
+
 		return goja.Undefined()
 	}
 }
@@ -1218,6 +1228,10 @@ func (im *RuntimeJavascriptInitModule) registerRtAfter(r *goja.Runtime) func(goj
 		lKey := strings.ToLower(RTAPI_PREFIX + key)
 		im.registerCallbackFn(RuntimeExecutionModeAfter, lKey, fnKey)
 		im.announceCallbackFn(RuntimeExecutionModeAfter, lKey)
+
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
 
 		return goja.Undefined()
 	}
@@ -1292,6 +1306,10 @@ func (im *RuntimeJavascriptInitModule) registerMatchmakerMatched(r *goja.Runtime
 		im.registerCallbackFn(RuntimeExecutionModeMatchmaker, "", fnKey)
 		im.announceCallbackFn(RuntimeExecutionModeMatchmaker, "")
 
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
+
 		return goja.Undefined()
 	}
 }
@@ -1310,6 +1328,10 @@ func (im *RuntimeJavascriptInitModule) registerTournamentEnd(r *goja.Runtime) fu
 		}
 		im.registerCallbackFn(RuntimeExecutionModeTournamentEnd, "", fnKey)
 		im.announceCallbackFn(RuntimeExecutionModeTournamentEnd, "")
+
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
 
 		return goja.Undefined()
 	}
@@ -1330,6 +1352,10 @@ func (im *RuntimeJavascriptInitModule) registerTournamentReset(r *goja.Runtime) 
 		im.registerCallbackFn(RuntimeExecutionModeTournamentReset, "", fnKey)
 		im.announceCallbackFn(RuntimeExecutionModeTournamentReset, "")
 
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
+
 		return goja.Undefined()
 	}
 }
@@ -1348,6 +1374,33 @@ func (im *RuntimeJavascriptInitModule) registerLeaderboardReset(r *goja.Runtime)
 		}
 		im.registerCallbackFn(RuntimeExecutionModeLeaderboardReset, "", fnKey)
 		im.announceCallbackFn(RuntimeExecutionModeLeaderboardReset, "")
+
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
+
+		return goja.Undefined()
+	}
+}
+
+func (im *RuntimeJavascriptInitModule) registerShutdown(r *goja.Runtime) func(goja.FunctionCall) goja.Value {
+	return func(f goja.FunctionCall) goja.Value {
+		fn := f.Argument(0)
+		_, ok := goja.AssertFunction(fn)
+		if !ok {
+			panic(r.NewTypeError("expects a function"))
+		}
+
+		fnKey, err := im.extractHookFn("registerShutdown")
+		if err != nil {
+			panic(r.NewGoError(err))
+		}
+		im.registerCallbackFn(RuntimeExecutionModeShutdown, "", fnKey)
+		im.announceCallbackFn(RuntimeExecutionModeShutdown, "")
+
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
 
 		return goja.Undefined()
 	}
@@ -1368,6 +1421,10 @@ func (im *RuntimeJavascriptInitModule) registerPurchaseNotificationApple(r *goja
 		im.registerCallbackFn(RuntimeExecutionModePurchaseNotificationApple, "", fnKey)
 		im.announceCallbackFn(RuntimeExecutionModePurchaseNotificationApple, "")
 
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
+
 		return goja.Undefined()
 	}
 }
@@ -1386,6 +1443,10 @@ func (im *RuntimeJavascriptInitModule) registerSubscriptionNotificationApple(r *
 		}
 		im.registerCallbackFn(RuntimeExecutionModeSubscriptionNotificationApple, "", fnKey)
 		im.announceCallbackFn(RuntimeExecutionModeSubscriptionNotificationApple, "")
+
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
 
 		return goja.Undefined()
 	}
@@ -1406,6 +1467,10 @@ func (im *RuntimeJavascriptInitModule) registerPurchaseNotificationGoogle(r *goj
 		im.registerCallbackFn(RuntimeExecutionModePurchaseNotificationGoogle, "", fnKey)
 		im.announceCallbackFn(RuntimeExecutionModePurchaseNotificationGoogle, "")
 
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
+
 		return goja.Undefined()
 	}
 }
@@ -1424,6 +1489,10 @@ func (im *RuntimeJavascriptInitModule) registerSubscriptionNotificationGoogle(r 
 		}
 		im.registerCallbackFn(RuntimeExecutionModeSubscriptionNotificationGoogle, "", fnKey)
 		im.announceCallbackFn(RuntimeExecutionModeSubscriptionNotificationGoogle, "")
+
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
 
 		return goja.Undefined()
 	}
@@ -1459,6 +1528,10 @@ func (im *RuntimeJavascriptInitModule) registerMatch(r *goja.Runtime) func(goja.
 		}
 		functions.initFn = fnKey
 
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
+
 		fnValue, ok = funcMap[string(MatchJoinAttempt)]
 		if !ok {
 			panic(r.NewTypeError(string(MatchJoinAttempt) + " not found"))
@@ -1472,6 +1545,10 @@ func (im *RuntimeJavascriptInitModule) registerMatch(r *goja.Runtime) func(goja.
 			panic(r.NewGoError(err))
 		}
 		functions.joinAttemptFn = fnKey
+
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
 
 		fnValue, ok = funcMap[string(MatchJoin)]
 		if !ok {
@@ -1487,6 +1564,10 @@ func (im *RuntimeJavascriptInitModule) registerMatch(r *goja.Runtime) func(goja.
 		}
 		functions.joinFn = fnKey
 
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
+
 		fnValue, ok = funcMap[string(MatchLeave)]
 		if !ok {
 			panic(r.NewTypeError(string(MatchLeave) + " not found"))
@@ -1500,6 +1581,10 @@ func (im *RuntimeJavascriptInitModule) registerMatch(r *goja.Runtime) func(goja.
 			panic(r.NewGoError(err))
 		}
 		functions.leaveFn = fnKey
+
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
 
 		fnValue, ok = funcMap[string(MatchLoop)]
 		if !ok {
@@ -1515,6 +1600,10 @@ func (im *RuntimeJavascriptInitModule) registerMatch(r *goja.Runtime) func(goja.
 		}
 		functions.loopFn = fnKey
 
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
+
 		fnValue, ok = funcMap[string(MatchTerminate)]
 		if !ok {
 			panic(r.NewTypeError(string(MatchTerminate) + " not found"))
@@ -1529,6 +1618,10 @@ func (im *RuntimeJavascriptInitModule) registerMatch(r *goja.Runtime) func(goja.
 		}
 		functions.terminateFn = fnKey
 
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
+
 		fnValue, ok = funcMap[string(MatchSignal)]
 		if !ok {
 			panic(r.NewTypeError(string(MatchSignal) + " not found"))
@@ -1542,6 +1635,10 @@ func (im *RuntimeJavascriptInitModule) registerMatch(r *goja.Runtime) func(goja.
 			panic(r.NewGoError(err))
 		}
 		functions.signalFn = fnKey
+
+		if err = im.checkFnScope(r, fnKey); err != nil {
+			panic(r.NewGoError(err))
+		}
 
 		im.MatchCallbacks.Add(name, functions)
 
@@ -1637,6 +1734,13 @@ func (im *RuntimeJavascriptInitModule) getMatchHookFnIdentifier(r *goja.Runtime,
 	return "", errors.New("not found")
 }
 
+func (im *RuntimeJavascriptInitModule) checkFnScope(r *goja.Runtime, key string) error {
+	if r.GlobalObject().Get(key) == nil {
+		return fmt.Errorf("function %q not registered in the global object scope", key)
+	}
+	return nil
+}
+
 func (im *RuntimeJavascriptInitModule) registerCallbackFn(mode RuntimeExecutionMode, key string, fn string) {
 	switch mode {
 	case RuntimeExecutionModeRPC:
@@ -1653,6 +1757,8 @@ func (im *RuntimeJavascriptInitModule) registerCallbackFn(mode RuntimeExecutionM
 		im.Callbacks.TournamentReset = fn
 	case RuntimeExecutionModeLeaderboardReset:
 		im.Callbacks.LeaderboardReset = fn
+	case RuntimeExecutionModeShutdown:
+		im.Callbacks.Shutdown = fn
 	case RuntimeExecutionModePurchaseNotificationApple:
 		im.Callbacks.PurchaseNotificationApple = fn
 	case RuntimeExecutionModeSubscriptionNotificationApple: